refactor(summary): relocate single-user types to their LCA
Rule 2 (a type lives under the lowest module that uses it) was the one
placement rule the earlier flat split left unaudited. Three types had
exactly one using module yet sat as its sibling:

- GasChange and SizeChange are the two halves of OutputVerdict::Changed and
  are touched nowhere else; move them under output_verdict/.
- ListingSection is SummaryTemplate.mover_sections and is built and held only
  there; move it under summary_template/.

Each moved module is wired through its parent with self:: imports, matching
the existing own-children convention. output_verdict.rs and summary_template.rs
become directory modules; the askama template path resolves against the crate
templates dir, so rendering is unaffected (fixtures still pass).
