+++ solx-benchmark-converter/src/output/summary/mod.rs
@@ -22,13 +22,10 @@ pub mod diff_counter;
 pub mod failure_regression;
 pub mod failure_regressions;
 pub mod failure_verdict;
-pub mod gas_change;
 pub mod health_issue;
-pub mod listing_section;
 pub mod movement;
 pub mod output_verdict;
 pub mod paired_bytes;
-pub mod size_change;
 pub mod suite_failures;
 pub mod suite_row;
 pub mod suite_stats;
